New Minecraft Versions supported (1.20.5 to 1.21.4), new Chat UI, new art assets (CC-BY-NC-SA-4.0)

+++ src/main/java/com/owlmaddie/goals/LeadPlayerGoal.java
@@ -1,6 +1,6 @@
 // SPDX-FileCopyrightText: 2025 owlmaddie LLC
 // SPDX-License-Identifier: GPL-3.0-or-later
-// Assets CC BY-NC 4.0; CreatureChat™ trademark © owlmaddie LLC - unauthorized use prohibited
+// Assets CC-BY-NC-SA-4.0; CreatureChat™ trademark © owlmaddie LLC - unauthorized use prohibited
 package com.owlmaddie.goals;
 
 import com.owlmaddie.chat.ChatDataManager;
@@ -145,7 +145,7 @@ public class LeadPlayerGoal extends PlayerBaseGoal {
             ServerWorld serverWorld = (ServerWorld) this.entity.getWorld();
 
             // Pass the angle using the "speed" argument, with deltaX, deltaY, deltaZ set to 0
-            LeadParticleEffect effect = new LeadParticleEffect(angle);
+            LeadParticleEffect effect = new LeadParticleEffect((float)angle);
             serverWorld.spawnParticles(effect, position.x, position.y + 0.05, position.z, 1, 0, 0, 0, 0);
         }
     }
